[z2432] Avoid last Visual Studio warning.
(based on commit [12596] - 9f426bd)


(based on commit [s2087] - 5595989)

+++ dep/src/g3dlite/g3dmath.cpp
@@ -41,7 +41,7 @@ double inf() {
 }
 
 bool isNaN(float x) {
-    static const float n = nan();
+    static const float n = static_cast<float>(nan());
     return memcmp(&x, &n, sizeof(float)) == 0;
 }
